Issue #2929577 by jwilson3: Add 'Is_Node' block condition plugin
This block condition helps the main content block placement in the IU theme
not be dependent on specific content types, but on all of them. Usage should
be coupled with enabling the IU Page Layout for all content types and placing
the Section field in an 'edge-to-edge' region, and the body field in the
content (default) region.